refactor(demo): refactor DependencyList into its own file
